Refactor Maven profiles and upgrade Scala to 2.13
- [Scala] Migrate deprecated JavaConversions to CollectionConverters for Scala 2.13
- [Scala] Fix explicit collection conversions in AgentManager and AccountService
- [Build] Synchronize Maven profile IDs between parent and sub-modules
- [Build] Use conditional module inclusion for Java 21 instead of JAVA_21_HOME
- [Java 21] Add add-exports compiler arguments for sun.net.www internal API access
- [Deploy] Support flexible assembly in scouter-deploy via Ant fileset untar
- [Misc] Disable failOnError for Javadoc to support mixed Java and Scala source
