+++ scouter.agent.java/pom.xml
@@ -18,11 +18,11 @@
         <profile>
             <id>java-8-plus</id>
             <activation>
+                <jdk>1.8</jdk>
                 <activeByDefault>true</activeByDefault>
             </activation>
             <properties>
                 <env>java8plus</env>
-                <java.version>1.8</java.version>
                 <scouter.agent.java.assembly.name.p>${scouter.agent.java.assembly.name}</scouter.agent.java.assembly.name.p>
             </properties>
             <dependencies>
@@ -31,30 +31,15 @@
                     <artifactId>scouter-extra-java20dummy</artifactId>
                 </dependency>
             </dependencies>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-compiler-plugin</artifactId>
-                        <configuration>
-                            <compilerVersion>${java.version}</compilerVersion>
-                            <source>${java.version}</source>
-                            <target>${java.version}</target>
-                            <fork>true</fork>
-                        </configuration>
-                    </plugin>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-javadoc-plugin</artifactId>
-                    </plugin>
-                </plugins>
-            </build>
         </profile>
+
         <profile>
             <id>java-21-plus</id>
+            <activation>
+                <jdk>[21,)</jdk>
+            </activation>
             <properties>
                 <env>java21plus</env>
-                <java.version>21</java.version>
                 <scouter.agent.java.assembly.name.p>${scouter.agent.java.assembly.name}.java21</scouter.agent.java.assembly.name.p>
             </properties>
             <dependencies>
@@ -69,24 +54,12 @@
                         <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-compiler-plugin</artifactId>
                         <configuration>
-                            <compilerVersion>${java.version}</compilerVersion>
-                            <source>${java.version}</source>
-                            <target>${java.version}</target>
-                            <fork>true</fork>
-                            <executable>${env.JAVA_21_HOME}/bin/javac</executable>
                             <compilerArgs>
                                 <arg>--add-exports</arg>
                                 <arg>java.base/sun.net.www.http=ALL-UNNAMED</arg>
                                 <arg>--add-exports</arg>
                                 <arg>java.base/sun.net.www.protocol.http=ALL-UNNAMED</arg>
                             </compilerArgs>
-                        </configuration>
-                    </plugin>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-javadoc-plugin</artifactId>
-                        <configuration>
-                            <javadocExecutable>${env.JAVA_21_HOME}/bin/javadoc</javadocExecutable>
                         </configuration>
                     </plugin>
                 </plugins>
@@ -140,19 +113,6 @@
                         <exclude>com/**</exclude>
                     </excludes>
                 </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <compilerVersion>${java.version}</compilerVersion>
-                    <source>${java.version}</source>
-                    <target>${java.version}</target>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-javadoc-plugin</artifactId>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -324,7 +284,6 @@
                                             value="${scouter.agent.file.name}" />
                                     </manifest>
                                 </jar>
-
                             </target>
                         </configuration>
                     </execution>
