+++ scouter.server/src/main/scala/scouter/server/term/handler/REALTIME.scala
@@ -29,7 +29,7 @@ import scouter.server.util.ThreadScala
 import scouter.util.DateUtil
 import scouter.util.FormatUtil
 import scouter.util.StringUtil
-import scala.collection.JavaConversions._
+import scala.jdk.CollectionConverters._
 object REALTIME {
 
     def process(cmd: String): Unit = {
